--- src/main/java/opensource/DreamingLibrary/example/group/dto/response/GroupResponse.java
@@ -1,21 +1,0 @@
-package opensource.DreamingLibrary.example.group.dto.response;
-
-import lombok.Builder;
-
-import java.time.LocalDateTime;
-
-/*
-1.
-
-
- */
-
-
-@Builder
-public record GroupResponse(
-        Long groupId,
-        String groupName,
-        LocalDateTime createdTime,
-        LocalDateTime updatedTime
-) {
-}
