--- src/main/java/opensource/DreamingLibrary/example/group/repository/GroupRepository.java
@@ -1,9 +1,0 @@
-package opensource.DreamingLibrary.example.group.repository;
-
-import opensource.DreamingLibrary.example.group.entity.Group;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface GroupRepository extends JpaRepository<Group, Long> {
-
-}
