--- src/main/java/opensource/DreamingLibrary/example/user/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package opensource.DreamingLibrary.example.user.repository;
-
-import opensource.DreamingLibrary.example.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Boolean existsByLoginId(String loginId);
-}
