+++ src/main/java/opensource/DreamingLibrary/group/mapper/GroupMapper.java
@@ -1,0 +1,17 @@
+package opensource.DreamingLibrary.group.mapper;
+
+import opensource.DreamingLibrary.group.dto.request.GroupCreateRequest;
+import opensource.DreamingLibrary.group.entity.Group;
+
+public class GroupMapper {
+
+    // DTO -> 엔티티 변환 (static 메서드)
+    public static Group from(GroupCreateRequest request) {
+        return Group.builder()
+                .groupName(request.groupName())
+                .createdTime(request.createdTime())
+                .updatedTime(request.updatedTime())
+                .build();
+    }
+
+}
