+++ src/main/java/opensource/DreamingLibrary/group/repository/GroupRepository.java
@@ -1,0 +1,9 @@
+package opensource.DreamingLibrary.group.repository;
+
+import opensource.DreamingLibrary.group.entity.Group;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface GroupRepository extends JpaRepository<Group, Long> {
+
+}
