+++ src/main/java/opensource/DreamingLibrary/group/service/GroupService.java
@@ -1,0 +1,48 @@
+package opensource.DreamingLibrary.group.service;
+
+
+import lombok.RequiredArgsConstructor;
+import opensource.DreamingLibrary.group.dto.request.GroupCreateRequest;
+import opensource.DreamingLibrary.group.dto.response.GroupResponse;
+import opensource.DreamingLibrary.group.entity.Group;
+import opensource.DreamingLibrary.group.repository.GroupRepository;
+import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class GroupService {
+
+    private final GroupRepository groupRepository;
+
+    // CREATE 예시
+    public GroupResponse createGroup(GroupCreateRequest requestDto) {
+        Group group = Group.builder()
+                .groupName(requestDto.groupName())
+                .createdTime(requestDto.createdTime() != null ? requestDto.createdTime() : LocalDateTime.now())
+                .updatedTime(requestDto.updatedTime() != null ? requestDto.updatedTime() : LocalDateTime.now())
+                .build();
+
+        Group saved = groupRepository.save(group);
+        return toResponse(saved);
+    }
+
+    // READ 예시
+    public GroupResponse getGroup(Long groupId) {
+        Optional<Group> optionalGroup = groupRepository.findById(groupId);
+        return optionalGroup
+                .map(this::toResponse)
+                .orElse(null); // 실제로는 예외처리
+    }
+
+    private GroupResponse toResponse(Group group) {
+        return GroupResponse.builder()
+                .groupId(group.getGroupId())
+                .groupName(group.getGroupName())
+                .createdTime(group.getCreatedTime())
+                .updatedTime(group.getUpdatedTime())
+                .build();
+    }
+}
