+++ src/main/java/opensource/DreamingLibrary/rent/repository/RentRepository.java
@@ -1,0 +1,7 @@
+package opensource.DreamingLibrary.rent.repository;
+
+import opensource.DreamingLibrary.rent.entity.Rent;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RentRepository extends JpaRepository<Rent, Long> {
+}
