+++ src/main/java/opensource/DreamingLibrary/user/entity/User.java
@@ -1,0 +1,40 @@
+package opensource.DreamingLibrary.user.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Entity
+@NoArgsConstructor
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotBlank
+    private String loginId;
+
+    @NotBlank
+    private String password;
+
+    @NotBlank
+    private String name;
+
+    @Builder
+    public User(
+            Long id,
+            String loginId,
+            String password,
+            String name){
+        this.id = id;
+        this.loginId = loginId;
+        this.password = password;
+        this.name = name;
+    }
+}
