+++ src/main/java/opensource/DreamingLibrary/user/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package opensource.DreamingLibrary.user.repository;
+
+import opensource.DreamingLibrary.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Boolean existsByLoginId(String loginId);
+}
